Add new store shorthand
