+++ tsdocs.config.js
@@ -2,4 +2,6 @@
 module.exports = {
   customPages: "./.docs",
   tsconfig: "./.docs/tsconfig.json",
+  stripInternal: true,
+  changelog: "CHANGELOG.md",
 }
